+++ tfhe/src/integer/gpu/server_key/radix/tests_unsigned/test_add.rs
@@ -1,0 +1,37 @@
+use crate::integer::gpu::server_key::radix::tests_unsigned::{
+    create_gpu_parametrized_test, GpuFunctionExecutor,
+};
+use crate::integer::gpu::CudaServerKey;
+use crate::integer::server_key::radix_parallel::tests_cases_unsigned::{
+    default_add_test, unchecked_add_assign_test, unchecked_add_test,
+};
+use crate::shortint::parameters::*;
+use crate::shortint::PBSParameters;
+
+create_gpu_parametrized_test!(integer_unchecked_add);
+create_gpu_parametrized_test!(integer_unchecked_add_assign);
+create_gpu_parametrized_test!(integer_add);
+
+fn integer_unchecked_add<P>(param: P)
+where
+    P: Into<PBSParameters>,
+{
+    let executor = GpuFunctionExecutor::new(&CudaServerKey::unchecked_add);
+    unchecked_add_test(param, executor);
+}
+
+fn integer_unchecked_add_assign<P>(param: P)
+where
+    P: Into<PBSParameters>,
+{
+    let executor = GpuFunctionExecutor::new(&CudaServerKey::unchecked_add_assign);
+    unchecked_add_assign_test(param, executor);
+}
+
+fn integer_add<P>(param: P)
+where
+    P: Into<PBSParameters>,
+{
+    let executor = GpuFunctionExecutor::new(&CudaServerKey::add);
+    default_add_test(param, executor);
+}
