+++ tfhe/src/integer/gpu/server_key/radix/tests_unsigned/test_sub.rs
@@ -1,0 +1,28 @@
+use crate::integer::gpu::server_key::radix::tests_unsigned::{
+    create_gpu_parametrized_test, GpuFunctionExecutor,
+};
+use crate::integer::gpu::CudaServerKey;
+use crate::integer::server_key::radix_parallel::tests_cases_unsigned::{
+    default_sub_test, unchecked_sub_test,
+};
+use crate::shortint::parameters::*;
+use crate::shortint::PBSParameters;
+
+create_gpu_parametrized_test!(integer_unchecked_sub);
+create_gpu_parametrized_test!(integer_sub);
+
+fn integer_unchecked_sub<P>(param: P)
+where
+    P: Into<PBSParameters>,
+{
+    let executor = GpuFunctionExecutor::new(&CudaServerKey::unchecked_sub);
+    unchecked_sub_test(param, executor);
+}
+
+fn integer_sub<P>(param: P)
+where
+    P: Into<PBSParameters>,
+{
+    let executor = GpuFunctionExecutor::new(&CudaServerKey::sub);
+    default_sub_test(param, executor);
+}
