+++ tfhe/src/integer/server_key/radix_parallel/mod.rs
@@ -27,9 +27,10 @@ pub(crate) mod tests_cases_signed;
 #[cfg(test)]
 pub(crate) mod tests_cases_unsigned;
 #[cfg(test)]
-mod tests_signed;
+pub(crate) mod tests_signed;
 #[cfg(test)]
 pub(crate) mod tests_unsigned;
+
 use crate::integer::ciphertext::IntegerRadixCiphertext;
 
 use super::ServerKey;
