+++ tfhe/src/integer/server_key/radix_parallel/tests_signed/mod.rs
@@ -1,6 +1,10 @@
+pub(crate) mod test_add;
+pub(crate) mod test_neg;
+pub(crate) mod test_sub;
+
 use crate::integer::keycache::KEY_CACHE;
-use crate::integer::server_key::radix_parallel::sub::SignedOperation;
 use crate::integer::server_key::radix_parallel::tests_cases_signed::*;
+use crate::integer::server_key::radix_parallel::tests_cases_unsigned::FunctionExecutor;
 use crate::integer::server_key::radix_parallel::tests_unsigned::CpuFunctionExecutor;
 use crate::integer::{
     BooleanBlock, IntegerKeyKind, RadixClientKey, ServerKey, SignedRadixCiphertext,
@@ -12,6 +16,7 @@ use crate::shortint::parameters::*;
 use itertools::iproduct;
 use paste::paste;
 use rand::Rng;
+use std::sync::Arc;
 
 /// Number of loop iteration within randomized tests
 #[cfg(not(tarpaulin))]
@@ -105,6 +110,67 @@ macro_rules! create_parametrized_test{
     };
 }
 
+impl<'a, F> FunctionExecutor<&'a SignedRadixCiphertext, (SignedRadixCiphertext, BooleanBlock)>
+    for CpuFunctionExecutor<F>
+where
+    F: Fn(&ServerKey, &SignedRadixCiphertext) -> (SignedRadixCiphertext, BooleanBlock),
+{
+    fn setup(&mut self, _cks: &RadixClientKey, sks: Arc<ServerKey>) {
+        self.sks = Some(sks);
+    }
+
+    fn execute(
+        &mut self,
+        input: &'a SignedRadixCiphertext,
+    ) -> (SignedRadixCiphertext, BooleanBlock) {
+        let sks = self.sks.as_ref().expect("setup was not properly called");
+        (self.func)(sks, input)
+    }
+}
+
+impl<'a, F> FunctionExecutor<&'a SignedRadixCiphertext, SignedRadixCiphertext>
+    for CpuFunctionExecutor<F>
+where
+    F: Fn(&ServerKey, &SignedRadixCiphertext) -> SignedRadixCiphertext,
+{
+    fn setup(&mut self, _cks: &RadixClientKey, sks: Arc<ServerKey>) {
+        self.sks = Some(sks);
+    }
+
+    fn execute(&mut self, input: &'a SignedRadixCiphertext) -> SignedRadixCiphertext {
+        let sks = self.sks.as_ref().expect("setup was not properly called");
+        (self.func)(sks, input)
+    }
+}
+impl<'a, F> FunctionExecutor<&'a mut SignedRadixCiphertext, ()> for CpuFunctionExecutor<F>
+where
+    F: Fn(&ServerKey, &'a mut SignedRadixCiphertext),
+{
+    fn setup(&mut self, _cks: &RadixClientKey, sks: Arc<ServerKey>) {
+        self.sks = Some(sks);
+    }
+
+    fn execute(&mut self, input: &'a mut SignedRadixCiphertext) {
+        let sks = self.sks.as_ref().expect("setup was not properly called");
+        (self.func)(sks, input);
+    }
+}
+
+impl<'a, F> FunctionExecutor<&'a mut SignedRadixCiphertext, SignedRadixCiphertext>
+    for CpuFunctionExecutor<F>
+where
+    F: Fn(&ServerKey, &mut SignedRadixCiphertext) -> SignedRadixCiphertext,
+{
+    fn setup(&mut self, _cks: &RadixClientKey, sks: Arc<ServerKey>) {
+        self.sks = Some(sks);
+    }
+
+    fn execute(&mut self, input: &'a mut SignedRadixCiphertext) -> SignedRadixCiphertext {
+        let sks = self.sks.as_ref().expect("setup was not properly called");
+        (self.func)(sks, input)
+    }
+}
+
 //================================================================================
 //     Encrypt/Decrypt Tests
 //================================================================================
@@ -165,49 +231,6 @@ fn integer_signed_encrypt_decrypt(param: impl Into<PBSParameters>) {
 //     Unchecked Tests
 //================================================================================
 
-create_parametrized_test!(integer_signed_unchecked_add);
-create_parametrized_test!(integer_signed_unchecked_overflowing_add);
-create_parametrized_test!(
-    integer_signed_unchecked_overflowing_add_parallelized {
-        coverage => {
-            COVERAGE_PARAM_MESSAGE_2_CARRY_2_KS_PBS,
-            COVERAGE_PARAM_MULTI_BIT_MESSAGE_2_CARRY_2_GROUP_2_KS_PBS,
-        },
-        no_coverage => {
-            // Requires 4 bits, so 1_1 parameters are not supported
-            // until they get their own version of the algorithm
-            PARAM_MESSAGE_2_CARRY_2_KS_PBS,
-            PARAM_MESSAGE_3_CARRY_3_KS_PBS,
-            PARAM_MESSAGE_4_CARRY_4_KS_PBS,
-            PARAM_MULTI_BIT_MESSAGE_2_CARRY_2_GROUP_2_KS_PBS,
-            PARAM_MULTI_BIT_MESSAGE_3_CARRY_3_GROUP_2_KS_PBS,
-            PARAM_MULTI_BIT_MESSAGE_2_CARRY_2_GROUP_3_KS_PBS,
-            PARAM_MULTI_BIT_MESSAGE_3_CARRY_3_GROUP_3_KS_PBS,
-        }
-    }
-);
-create_parametrized_test!(integer_signed_unchecked_neg);
-create_parametrized_test!(integer_signed_unchecked_sub);
-create_parametrized_test!(integer_signed_unchecked_overflowing_sub);
-create_parametrized_test!(
-    integer_signed_unchecked_overflowing_sub_parallelized {
-        coverage => {
-            COVERAGE_PARAM_MESSAGE_2_CARRY_2_KS_PBS,
-            COVERAGE_PARAM_MULTI_BIT_MESSAGE_2_CARRY_2_GROUP_2_KS_PBS,
-        },
-        no_coverage => {
-            // Requires 4 bits, so 1_1 parameters are not supported
-            // until they get their own version of the algorithm
-            PARAM_MESSAGE_2_CARRY_2_KS_PBS,
-            PARAM_MESSAGE_3_CARRY_3_KS_PBS,
-            PARAM_MESSAGE_4_CARRY_4_KS_PBS,
-            PARAM_MULTI_BIT_MESSAGE_2_CARRY_2_GROUP_2_KS_PBS,
-            PARAM_MULTI_BIT_MESSAGE_3_CARRY_3_GROUP_2_KS_PBS,
-            PARAM_MULTI_BIT_MESSAGE_2_CARRY_2_GROUP_3_KS_PBS,
-            PARAM_MULTI_BIT_MESSAGE_3_CARRY_3_GROUP_3_KS_PBS,
-        }
-    }
-);
 create_parametrized_test!(integer_signed_unchecked_mul);
 create_parametrized_test!(integer_signed_unchecked_bitand);
 create_parametrized_test!(integer_signed_unchecked_bitor);
@@ -329,376 +352,6 @@ create_parametrized_test!(
     }
 );
 create_parametrized_test!(integer_signed_unchecked_absolute_value);
-
-fn integer_signed_unchecked_add<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let executor = CpuFunctionExecutor::new(&ServerKey::unchecked_add_parallelized);
-    signed_unchecked_add_test(param, executor);
-}
-
-fn signed_unchecked_overflowing_add_test_case<P, F>(param: P, signed_overflowing_add: F)
-where
-    P: Into<PBSParameters>,
-    F: for<'a> Fn(
-        &'a ServerKey,
-        &'a SignedRadixCiphertext,
-        &'a SignedRadixCiphertext,
-    ) -> (SignedRadixCiphertext, BooleanBlock),
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    sks.set_deterministic_pbs_execution(true);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-    let hardcoded_values = [
-        (-modulus, -1),
-        (modulus - 1, 1),
-        (-1, -modulus),
-        (1, modulus - 1),
-    ];
-    for (clear_0, clear_1) in hardcoded_values {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let (ct_res, result_overflowed) =
-            sks.signed_overflowing_scalar_add_parallelized(&ctxt_0, clear_1);
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_add_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-        let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for add, for ({clear_0} + {clear_1}) % {modulus} \
-             expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_add for ({clear_0} + {clear_1}) % {modulus} \
-             expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert_eq!(result_overflowed.0.degree.get(), 1);
-        assert_eq!(result_overflowed.0.noise_level(), NoiseLevel::NOMINAL);
-    }
-
-    for _ in 0..NB_TESTS {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let (ct_res, result_overflowed) = signed_overflowing_add(&sks, &ctxt_0, &ctxt_1);
-        let (tmp_ct, tmp_o) = signed_overflowing_add(&sks, &ctxt_0, &ctxt_1);
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp_ct, "Failed determinism check");
-        assert_eq!(tmp_o, result_overflowed, "Failed determinism check");
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_add_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-        let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for add, for ({clear_0} + {clear_1}) % {modulus} \
-             expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_add for ({clear_0} + {clear_1}) % {modulus} \
-             expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-    }
-
-    // Test with trivial inputs, as it was bugged at some point
-    let values = [
-        (rng.gen::<i64>() % modulus, 0i64),
-        (rng.gen::<i64>() % modulus, rng.gen::<i64>() % modulus),
-        (rng.gen::<i64>() % modulus, rng.gen::<i64>() % modulus),
-        (rng.gen::<i64>() % modulus, rng.gen::<i64>() % modulus),
-        (rng.gen::<i64>() % modulus, rng.gen::<i64>() % modulus),
-    ];
-    for (clear_0, clear_1) in values {
-        let a: SignedRadixCiphertext = sks.create_trivial_radix(clear_0, NB_CTXT);
-        let b: SignedRadixCiphertext = sks.create_trivial_radix(clear_1, NB_CTXT);
-
-        let (encrypted_result, encrypted_overflow) = signed_overflowing_add(&sks, &a, &b);
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_add_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&encrypted_result);
-        let decrypted_overflowed = cks.decrypt_bool(&encrypted_overflow);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for add, for ({clear_0} + {clear_1}) % {modulus} \
-                expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_add, for ({clear_0} + {clear_1}) % {modulus} \
-                expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-    }
-}
-
-fn integer_signed_unchecked_overflowing_add<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    // Calls the low level function like this so we are sure the sequential version is tested
-    let func = |sks: &ServerKey,
-                lhs: &SignedRadixCiphertext,
-                rhs: &SignedRadixCiphertext|
-     -> (SignedRadixCiphertext, BooleanBlock) {
-        sks.unchecked_signed_overflowing_add_or_sub(lhs, rhs, SignedOperation::Addition)
-    };
-    signed_unchecked_overflowing_add_test_case(param, func);
-}
-
-fn integer_signed_unchecked_overflowing_add_parallelized<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    // Calls the low level function like this so we are sure the parallel version is tested
-    //
-    // However this only supports param X_X where X >= 2
-    let func = |sks: &ServerKey,
-                lhs: &SignedRadixCiphertext,
-                rhs: &SignedRadixCiphertext|
-     -> (SignedRadixCiphertext, BooleanBlock) {
-        sks.unchecked_signed_overflowing_add_or_sub_parallelized_impl(
-            lhs,
-            rhs,
-            SignedOperation::Addition,
-        )
-    };
-    signed_unchecked_overflowing_add_test_case(param, func);
-}
-
-// There is no unchecked_neg_parallelized,
-// but test the non parallel version here anyway
-// as it is used in other parallel ops (e.g: sub)
-fn integer_signed_unchecked_neg(param: impl Into<PBSParameters>) {
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    // -modulus is a special case, its negation cannot be
-    // represented. rust by default returns -modulus
-    // (which is what two complement result in)
-    {
-        let clear = -modulus;
-        let ctxt = cks.encrypt_signed_radix(clear, NB_CTXT);
-
-        let ct_res = sks.neg_parallelized(&ctxt);
-
-        let dec: i64 = cks.decrypt_signed_radix(&ct_res);
-        let clear_result = signed_neg_under_modulus(clear, modulus);
-
-        assert_eq!(clear_result, dec);
-        assert_eq!(clear_result, -modulus);
-    }
-
-    for (clear_0, _) in
-        create_iterator_of_signed_random_pairs::<NB_TESTS_UNCHECKED>(&mut rng, modulus)
-    {
-        let ctxt_0 = cks.encrypt_signed_radix(clear_0, NB_CTXT);
-
-        let ct_res = sks.unchecked_neg(&ctxt_0);
-        let dec_res: i64 = cks.decrypt_signed_radix(&ct_res);
-        let clear_res = signed_neg_under_modulus(clear_0, modulus);
-        assert_eq!(clear_res, dec_res);
-    }
-
-    // negation of trivial 0
-    {
-        let ctxt_0 = sks.create_trivial_radix(0i64, NB_CTXT);
-        let ct_res = sks.unchecked_neg(&ctxt_0);
-        let dec_res: i64 = cks.decrypt_signed_radix(&ct_res);
-        assert_eq!(0, dec_res);
-    }
-}
-
-fn integer_signed_unchecked_sub(param: impl Into<PBSParameters>) {
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    // check some overflow behaviour
-    let overflowing_values = [
-        (-modulus, 1, modulus - 1),
-        (modulus - 1, -1, -modulus),
-        (-modulus, 2, modulus - 2),
-        (modulus - 2, -2, -modulus),
-    ];
-    for (clear_0, clear_1, expected_clear) in overflowing_values {
-        let ctxt_0 = cks.encrypt_signed_radix(clear_0, NB_CTXT);
-        let ctxt_1 = cks.encrypt_signed_radix(clear_1, NB_CTXT);
-        let ct_res = sks.unchecked_sub(&ctxt_0, &ctxt_1);
-        let dec_res: i64 = cks.decrypt_signed_radix(&ct_res);
-        let clear_res = signed_sub_under_modulus(clear_0, clear_1, modulus);
-        assert_eq!(clear_res, dec_res);
-        assert_eq!(clear_res, expected_clear);
-    }
-
-    for (clear_0, clear_1) in
-        create_iterator_of_signed_random_pairs::<NB_TESTS_UNCHECKED>(&mut rng, modulus)
-    {
-        let ctxt_0 = cks.encrypt_signed_radix(clear_0, NB_CTXT);
-        let ctxt_1 = cks.encrypt_signed_radix(clear_1, NB_CTXT);
-
-        let ct_res = sks.unchecked_sub(&ctxt_0, &ctxt_1);
-        let dec_res: i64 = cks.decrypt_signed_radix(&ct_res);
-        let clear_res = signed_sub_under_modulus(clear_0, clear_1, modulus);
-        assert_eq!(clear_res, dec_res);
-    }
-}
-
-fn signed_unchecked_overflowing_sub_test_case<P, F>(param: P, signed_overflowing_sub: F)
-where
-    P: Into<PBSParameters>,
-    F: for<'a> Fn(
-        &'a ServerKey,
-        &'a SignedRadixCiphertext,
-        &'a SignedRadixCiphertext,
-    ) -> (SignedRadixCiphertext, BooleanBlock),
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    sks.set_deterministic_pbs_execution(true);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    let hardcoded_values = [
-        (-modulus, 1),
-        (modulus - 1, -1),
-        (1, -modulus),
-        (-1, modulus - 1),
-    ];
-    for (clear_0, clear_1) in hardcoded_values {
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-
-        let (ct_res, result_overflowed) =
-            sks.signed_overflowing_scalar_sub_parallelized(&ctxt_0, clear_1);
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_sub_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-        let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for overflowing_sub, for ({clear_0} - {clear_1}) % {modulus} \
-             expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_sub for ({clear_0} - {clear_1}) % {modulus} \
-             expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert_eq!(result_overflowed.0.degree.get(), 1);
-        assert_eq!(result_overflowed.0.noise_level(), NoiseLevel::NOMINAL);
-    }
-
-    for _ in 0..NB_TESTS {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let (ct_res, result_overflowed) = signed_overflowing_sub(&sks, &ctxt_0, &ctxt_1);
-        let (tmp_ct, tmp_o) = signed_overflowing_sub(&sks, &ctxt_0, &ctxt_1);
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp_ct, "Failed determinism check");
-        assert_eq!(tmp_o, result_overflowed, "Failed determinism check");
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_sub_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-        let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for sub, for ({clear_0} - {clear_1}) % {modulus} \
-             expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_sub for ({clear_0} - {clear_1}) % {modulus} \
-             expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-    }
-
-    // Test with trivial inputs, as it was bugged at some point
-    let values = [
-        (rng.gen::<i64>() % modulus, 0i64),
-        (rng.gen::<i64>() % modulus, rng.gen::<i64>() % modulus),
-        (rng.gen::<i64>() % modulus, rng.gen::<i64>() % modulus),
-        (rng.gen::<i64>() % modulus, rng.gen::<i64>() % modulus),
-        (rng.gen::<i64>() % modulus, rng.gen::<i64>() % modulus),
-    ];
-    for (clear_0, clear_1) in values {
-        let a: SignedRadixCiphertext = sks.create_trivial_radix(clear_0, NB_CTXT);
-        let b: SignedRadixCiphertext = sks.create_trivial_radix(clear_1, NB_CTXT);
-
-        let (encrypted_result, encrypted_overflow) = signed_overflowing_sub(&sks, &a, &b);
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_sub_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&encrypted_result);
-        let decrypted_overflowed = cks.decrypt_bool(&encrypted_overflow);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for sub, for ({clear_0} - {clear_1}) % {modulus} \
-                expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_sub, for ({clear_0} - {clear_1}) % {modulus} \
-                expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-    }
-}
-
-fn integer_signed_unchecked_overflowing_sub<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    signed_unchecked_overflowing_sub_test_case(param, ServerKey::unchecked_signed_overflowing_sub);
-}
-
-fn integer_signed_unchecked_overflowing_sub_parallelized<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    // Call _impl so we are sure the parallel version is tested
-    //
-    // However this only supports param X_X where X >= 4
-    signed_unchecked_overflowing_sub_test_case(
-        param,
-        ServerKey::unchecked_signed_overflowing_sub_parallelized_impl,
-    );
-}
 
 fn integer_signed_unchecked_mul(param: impl Into<PBSParameters>) {
     let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
@@ -1162,75 +815,7 @@ fn integer_signed_unchecked_div_rem_floor(param: impl Into<PBSParameters>) {
 //     Smart Tests
 //================================================================================
 
-create_parametrized_test!(integer_signed_smart_add);
-create_parametrized_test!(integer_signed_smart_neg);
 create_parametrized_test!(integer_signed_smart_absolute_value);
-
-fn integer_signed_smart_add(param: impl Into<PBSParameters>) {
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    let mut clear;
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear_0 = rng.gen_range(-modulus..modulus);
-        let clear_1 = rng.gen_range(-modulus..modulus);
-
-        let mut ctxt_0 = cks.encrypt_signed_radix(clear_0, NB_CTXT);
-        let mut ctxt_1 = cks.encrypt_signed_radix(clear_1, NB_CTXT);
-
-        let mut ct_res = sks.smart_add_parallelized(&mut ctxt_0, &mut ctxt_1);
-        clear = signed_add_under_modulus(clear_0, clear_1, modulus);
-        let dec_res: i64 = cks.decrypt_signed_radix(&ct_res);
-        assert_eq!(clear, dec_res);
-
-        // add multiple times to raise the degree
-        for _ in 0..NB_TESTS_SMALLER {
-            ct_res = sks.smart_add_parallelized(&mut ct_res, &mut ctxt_0);
-            clear = signed_add_under_modulus(clear, clear_0, modulus);
-
-            let dec_res: i64 = cks.decrypt_signed_radix(&ct_res);
-
-            // println!("clear = {}, dec_res = {}", clear, dec_res);
-            assert_eq!(clear, dec_res);
-        }
-    }
-}
-
-fn integer_signed_smart_neg<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear = rng.gen::<i64>() % modulus;
-
-        let mut ctxt = cks.encrypt_signed(clear);
-
-        let mut ct_res = sks.smart_neg_parallelized(&mut ctxt);
-        let mut clear_res = signed_neg_under_modulus(clear, modulus);
-        let dec: i64 = cks.decrypt_signed(&ct_res);
-        assert_eq!(clear_res, dec);
-
-        for _ in 0..NB_TESTS_SMALLER {
-            ct_res = sks.smart_neg_parallelized(&mut ct_res);
-            clear_res = signed_neg_under_modulus(clear_res, modulus);
-
-            let dec: i64 = cks.decrypt_signed(&ct_res);
-            println!("clear_res: {clear_res}, dec : {dec}");
-            assert_eq!(clear_res, dec);
-        }
-    }
-}
 
 fn integer_signed_smart_absolute_value(param: impl Into<PBSParameters>) {
     let (cks, sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
@@ -1266,11 +851,6 @@ fn integer_signed_smart_absolute_value(param: impl Into<PBSParameters>) {
 //     Default Tests
 //================================================================================
 
-create_parametrized_test!(integer_signed_default_add);
-create_parametrized_test!(integer_signed_default_overflowing_add);
-create_parametrized_test!(integer_signed_default_neg);
-create_parametrized_test!(integer_signed_default_sub);
-create_parametrized_test!(integer_signed_default_overflowing_sub);
 create_parametrized_test!(integer_signed_default_mul);
 create_parametrized_test!(
     integer_signed_default_overflowing_mul {
@@ -1388,347 +968,6 @@ create_parametrized_test!(integer_signed_default_checked_ilog2 {
     PARAM_MULTI_BIT_MESSAGE_2_CARRY_2_GROUP_3_KS_PBS,
     PARAM_MULTI_BIT_MESSAGE_3_CARRY_3_GROUP_3_KS_PBS
 });
-
-fn integer_signed_default_add<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let executor = CpuFunctionExecutor::new(&ServerKey::add_parallelized);
-    signed_default_add_test(param, executor);
-}
-fn integer_signed_default_overflowing_add<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    sks.set_deterministic_pbs_execution(true);
-
-    let mut rng = rand::thread_rng();
-
-    // message_modulus^vec_length
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let (ct_res, result_overflowed) = sks.signed_overflowing_add_parallelized(&ctxt_0, &ctxt_1);
-        let (tmp_ct, tmp_o) = sks.signed_overflowing_add_parallelized(&ctxt_0, &ctxt_1);
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp_ct, "Failed determinism check");
-        assert_eq!(tmp_o, result_overflowed, "Failed determinism check");
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_add_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-        let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for add, for ({clear_0} + {clear_1}) % {modulus} \
-             expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_suv for ({clear_0} + {clear_1}) % {modulus} \
-             expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert_eq!(result_overflowed.0.degree.get(), 1);
-        assert_eq!(result_overflowed.0.noise_level(), NoiseLevel::NOMINAL);
-
-        for _ in 0..NB_TESTS_SMALLER {
-            // Add non zero scalar to have non clean ciphertexts
-            let clear_2 = random_non_zero_value(&mut rng, modulus);
-            let clear_3 = random_non_zero_value(&mut rng, modulus);
-
-            let ctxt_0 = sks.unchecked_scalar_add(&ctxt_0, clear_2);
-            let ctxt_1 = sks.unchecked_scalar_add(&ctxt_1, clear_3);
-
-            let clear_lhs = signed_add_under_modulus(clear_0, clear_2, modulus);
-            let clear_rhs = signed_add_under_modulus(clear_1, clear_3, modulus);
-
-            let d0: i64 = cks.decrypt_signed(&ctxt_0);
-            assert_eq!(d0, clear_lhs, "Failed sanity decryption check");
-            let d1: i64 = cks.decrypt_signed(&ctxt_1);
-            assert_eq!(d1, clear_rhs, "Failed sanity decryption check");
-
-            let (ct_res, result_overflowed) =
-                sks.signed_overflowing_add_parallelized(&ctxt_0, &ctxt_1);
-            assert!(ct_res.block_carries_are_empty());
-
-            let (expected_result, expected_overflowed) =
-                signed_overflowing_add_under_modulus(clear_lhs, clear_rhs, modulus);
-
-            let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-            let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-            assert_eq!(
-                decrypted_result, expected_result,
-                "Invalid result for add, for ({clear_lhs} + {clear_rhs}) % {modulus} \
-                expected {expected_result}, got {decrypted_result}"
-            );
-            assert_eq!(
-                decrypted_overflowed,
-                expected_overflowed,
-                "Invalid overflow flag result for overflowing_add, for ({clear_lhs} + {clear_rhs}) % {modulus} \
-                expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-            );
-            assert_eq!(result_overflowed.0.degree.get(), 1);
-            assert_eq!(result_overflowed.0.noise_level(), NoiseLevel::NOMINAL);
-        }
-    }
-
-    // Test with trivial inputs, as it was bugged at some point
-    for _ in 0..4 {
-        // Reduce maximum value of random number such that at least the last block is a trivial 0
-        // (This is how the reproducing case was found)
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let a: SignedRadixCiphertext = sks.create_trivial_radix(clear_0, NB_CTXT);
-        let b: SignedRadixCiphertext = sks.create_trivial_radix(clear_1, NB_CTXT);
-
-        let (encrypted_result, encrypted_overflow) =
-            sks.signed_overflowing_add_parallelized(&a, &b);
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_add_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&encrypted_result);
-        let decrypted_overflowed = cks.decrypt_bool(&encrypted_overflow);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for add, for ({clear_0} + {clear_1}) % {modulus} \
-                expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_add, for ({clear_0} + {clear_1}) % {modulus} \
-                expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert_eq!(encrypted_overflow.0.degree.get(), 1);
-        assert_eq!(encrypted_overflow.0.noise_level(), NoiseLevel::ZERO);
-    }
-}
-
-fn integer_signed_default_neg<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    sks.set_deterministic_pbs_execution(true);
-
-    let mut rng = rand::thread_rng();
-
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    // -modulus is a special case, its negation cannot be
-    // represented. rust by default returns -modulus
-    // (which is what two complement result in)
-    {
-        let clear = -modulus;
-        let ctxt = cks.encrypt_signed(clear);
-
-        let ct_res = sks.neg_parallelized(&ctxt);
-        let tmp = sks.neg_parallelized(&ctxt);
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp);
-
-        let dec: i64 = cks.decrypt_signed(&ct_res);
-        let clear_result = signed_neg_under_modulus(clear, modulus);
-
-        assert_eq!(clear_result, dec);
-    }
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear = rng.gen::<i64>() % modulus;
-
-        let ctxt = cks.encrypt_signed(clear);
-
-        let ct_res = sks.neg_parallelized(&ctxt);
-        let tmp = sks.neg_parallelized(&ctxt);
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp);
-
-        let dec: i64 = cks.decrypt_signed(&ct_res);
-        let clear_result = signed_neg_under_modulus(clear, modulus);
-
-        assert_eq!(clear_result, dec);
-    }
-}
-
-fn integer_signed_default_sub<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    sks.set_deterministic_pbs_execution(true);
-
-    let mut rng = rand::thread_rng();
-
-    // message_modulus^vec_length
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    let mut clear;
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let mut ct_res = sks.sub_parallelized(&ctxt_0, &ctxt_1);
-        let tmp_ct = sks.sub_parallelized(&ctxt_0, &ctxt_1);
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp_ct);
-
-        clear = signed_sub_under_modulus(clear_0, clear_1, modulus);
-
-        // sub multiple times to raise the degree
-        for _ in 0..NB_TESTS_SMALLER {
-            ct_res = sks.sub_parallelized(&ct_res, &ctxt_0);
-            assert!(ct_res.block_carries_are_empty());
-            clear = signed_sub_under_modulus(clear, clear_0, modulus);
-
-            let dec_res: i64 = cks.decrypt_signed(&ct_res);
-
-            // println!("clear = {}, dec_res = {}", clear, dec_res);
-            assert_eq!(clear, dec_res);
-        }
-    }
-}
-
-fn integer_signed_default_overflowing_sub<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let (cks, mut sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-    let cks = RadixClientKey::from((cks, NB_CTXT));
-
-    sks.set_deterministic_pbs_execution(true);
-
-    let mut rng = rand::thread_rng();
-
-    // message_modulus^vec_length
-    let modulus = (cks.parameters().message_modulus().0.pow(NB_CTXT as u32) / 2) as i64;
-
-    for _ in 0..NB_TESTS_SMALLER {
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let ctxt_0 = cks.encrypt_signed(clear_0);
-        let ctxt_1 = cks.encrypt_signed(clear_1);
-
-        let (ct_res, result_overflowed) = sks.signed_overflowing_sub_parallelized(&ctxt_0, &ctxt_1);
-        let (tmp_ct, tmp_o) = sks.signed_overflowing_sub_parallelized(&ctxt_0, &ctxt_1);
-        assert!(ct_res.block_carries_are_empty());
-        assert_eq!(ct_res, tmp_ct, "Failed determinism check");
-        assert_eq!(tmp_o, result_overflowed, "Failed determinism check");
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_sub_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-        let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for sub, for ({clear_0} - {clear_1}) % {modulus} \
-             expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_suv for ({clear_0} - {clear_1}) % {modulus} \
-             expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert_eq!(result_overflowed.0.degree.get(), 1);
-        assert_eq!(result_overflowed.0.noise_level(), NoiseLevel::NOMINAL);
-
-        for _ in 0..NB_TESTS_SMALLER {
-            // Add non zero scalar to have non clean ciphertexts
-            let clear_2 = random_non_zero_value(&mut rng, modulus);
-            let clear_3 = random_non_zero_value(&mut rng, modulus);
-
-            let ctxt_0 = sks.unchecked_scalar_add(&ctxt_0, clear_2);
-            let ctxt_1 = sks.unchecked_scalar_add(&ctxt_1, clear_3);
-
-            let clear_lhs = signed_add_under_modulus(clear_0, clear_2, modulus);
-            let clear_rhs = signed_add_under_modulus(clear_1, clear_3, modulus);
-
-            let d0: i64 = cks.decrypt_signed(&ctxt_0);
-            assert_eq!(d0, clear_lhs, "Failed sanity decryption check");
-            let d1: i64 = cks.decrypt_signed(&ctxt_1);
-            assert_eq!(d1, clear_rhs, "Failed sanity decryption check");
-
-            let (ct_res, result_overflowed) =
-                sks.signed_overflowing_sub_parallelized(&ctxt_0, &ctxt_1);
-            assert!(ct_res.block_carries_are_empty());
-
-            let (expected_result, expected_overflowed) =
-                signed_overflowing_sub_under_modulus(clear_lhs, clear_rhs, modulus);
-
-            let decrypted_result: i64 = cks.decrypt_signed(&ct_res);
-            let decrypted_overflowed = cks.decrypt_bool(&result_overflowed);
-            assert_eq!(
-                decrypted_result, expected_result,
-                "Invalid result for sub, for ({clear_lhs} - {clear_rhs}) % {modulus} \
-                expected {expected_result}, got {decrypted_result}"
-            );
-            assert_eq!(
-                decrypted_overflowed,
-                expected_overflowed,
-                "Invalid overflow flag result for overflowing_sub, for ({clear_lhs} - {clear_rhs}) % {modulus} \
-                expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-            );
-            assert_eq!(result_overflowed.0.degree.get(), 1);
-            assert_eq!(result_overflowed.0.noise_level(), NoiseLevel::NOMINAL);
-        }
-    }
-
-    // Test with trivial inputs, as it was bugged at some point
-    for _ in 0..4 {
-        // Reduce maximum value of random number such that at least the last block is a trivial 0
-        // (This is how the reproducing case was found)
-        let clear_0 = rng.gen::<i64>() % modulus;
-        let clear_1 = rng.gen::<i64>() % modulus;
-
-        let a: SignedRadixCiphertext = sks.create_trivial_radix(clear_0, NB_CTXT);
-        let b: SignedRadixCiphertext = sks.create_trivial_radix(clear_1, NB_CTXT);
-
-        let (encrypted_result, encrypted_overflow) =
-            sks.signed_overflowing_sub_parallelized(&a, &b);
-
-        let (expected_result, expected_overflowed) =
-            signed_overflowing_sub_under_modulus(clear_0, clear_1, modulus);
-
-        let decrypted_result: i64 = cks.decrypt_signed(&encrypted_result);
-        let decrypted_overflowed = cks.decrypt_bool(&encrypted_overflow);
-        assert_eq!(
-            decrypted_result, expected_result,
-            "Invalid result for sub, for ({clear_0} - {clear_1}) % {modulus} \
-                expected {expected_result}, got {decrypted_result}"
-        );
-        assert_eq!(
-            decrypted_overflowed,
-            expected_overflowed,
-            "Invalid overflow flag result for overflowing_sub, for ({clear_0} - {clear_1}) % {modulus} \
-                expected overflow flag {expected_overflowed}, got {decrypted_overflowed}"
-        );
-        assert_eq!(encrypted_overflow.0.degree.get(), 1);
-        assert_eq!(encrypted_overflow.0.noise_level(), NoiseLevel::ZERO);
-    }
-}
 
 fn integer_signed_default_mul<P>(param: P)
 where
