+++ tfhe/src/integer/server_key/radix_parallel/tests_unsigned/mod.rs
@@ -1,4 +1,6 @@
 pub(crate) mod test_add;
+pub(crate) mod test_neg;
+pub(crate) mod test_sub;
 
 use super::tests_cases_unsigned::*;
 use crate::integer::keycache::KEY_CACHE;
@@ -570,30 +572,6 @@ create_parametrized_test!(integer_unchecked_scalar_rotate_right);
 create_parametrized_test!(integer_unchecked_scalar_rotate_left);
 create_parametrized_test!(integer_default_scalar_rotate_right);
 create_parametrized_test!(integer_default_scalar_rotate_left);
-// negations
-create_parametrized_test!(integer_smart_neg);
-create_parametrized_test!(integer_default_neg);
-create_parametrized_test!(integer_smart_sub);
-create_parametrized_test!(integer_default_sub);
-create_parametrized_test!(integer_default_overflowing_sub);
-create_parametrized_test!(
-    integer_default_sub_work_efficient {
-        coverage => {
-            COVERAGE_PARAM_MESSAGE_2_CARRY_2_KS_PBS,
-            COVERAGE_PARAM_MULTI_BIT_MESSAGE_2_CARRY_2_GROUP_2_KS_PBS,
-        },
-        no_coverage => {
-            // This algorithm requires 3 bits
-            PARAM_MESSAGE_2_CARRY_2_KS_PBS,
-            PARAM_MESSAGE_3_CARRY_3_KS_PBS,
-            PARAM_MESSAGE_4_CARRY_4_KS_PBS,
-            PARAM_MULTI_BIT_MESSAGE_2_CARRY_2_GROUP_2_KS_PBS,
-            PARAM_MULTI_BIT_MESSAGE_3_CARRY_3_GROUP_2_KS_PBS,
-            PARAM_MULTI_BIT_MESSAGE_2_CARRY_2_GROUP_3_KS_PBS,
-            PARAM_MULTI_BIT_MESSAGE_3_CARRY_3_GROUP_3_KS_PBS,
-        }
-    }
-);
 create_parametrized_test!(integer_default_scalar_div_rem);
 create_parametrized_test!(integer_unchecked_block_mul);
 create_parametrized_test!(integer_smart_block_mul);
@@ -654,9 +632,9 @@ create_parametrized_test!(
 /// It will mainly simply forward call to a server key method
 pub(crate) struct CpuFunctionExecutor<F> {
     /// The server key is set later, when the test cast calls setup
-    sks: Option<Arc<ServerKey>>,
+    pub(crate) sks: Option<Arc<ServerKey>>,
     /// The server key function which will be called
-    func: F,
+    pub(crate) func: F,
 }
 
 impl<F> CpuFunctionExecutor<F> {
@@ -878,28 +856,12 @@ where
 // Smart Tests
 //=============================================================================
 
-fn integer_smart_sub<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let executor = CpuFunctionExecutor::new(&ServerKey::smart_sub_parallelized);
-    smart_sub_test(param, executor);
-}
-
 fn integer_smart_mul<P>(param: P)
 where
     P: Into<PBSParameters>,
 {
     let executor = CpuFunctionExecutor::new(&ServerKey::smart_mul_parallelized);
     smart_mul_test(param, executor);
-}
-
-fn integer_smart_neg<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let executor = CpuFunctionExecutor::new(&ServerKey::smart_neg_parallelized);
-    smart_neg_test(param, executor);
 }
 
 fn integer_smart_bitand<P>(param: P)
@@ -1091,30 +1053,6 @@ where
 // Default Tests
 //=============================================================================
 
-fn integer_default_sub<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let executor = CpuFunctionExecutor::new(&ServerKey::sub_parallelized);
-    default_sub_test(param, executor);
-}
-
-fn integer_default_overflowing_sub<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let executor = CpuFunctionExecutor::new(&ServerKey::unsigned_overflowing_sub_parallelized);
-    default_overflowing_sub_test(param, executor);
-}
-
-fn integer_default_sub_work_efficient<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let executor = CpuFunctionExecutor::new(&ServerKey::sub_parallelized_work_efficient);
-    default_sub_test(param, executor);
-}
-
 fn integer_default_mul<P>(param: P)
 where
     P: Into<PBSParameters>,
@@ -1129,14 +1067,6 @@ where
 {
     let executor = CpuFunctionExecutor::new(&ServerKey::unsigned_overflowing_mul_parallelized);
     default_overflowing_mul_test(param, executor);
-}
-
-fn integer_default_neg<P>(param: P)
-where
-    P: Into<PBSParameters>,
-{
-    let executor = CpuFunctionExecutor::new(&ServerKey::neg_parallelized);
-    default_neg_test(param, executor);
 }
 
 fn integer_default_bitand<P>(param: P)
