+++ tfhe/src/integer/server_key/radix_parallel/tests_unsigned/test_neg.rs
@@ -1,0 +1,28 @@
+use crate::integer::server_key::radix_parallel::tests_cases_unsigned::{
+    default_neg_test, smart_neg_test,
+};
+use crate::integer::server_key::radix_parallel::tests_unsigned::CpuFunctionExecutor;
+use crate::integer::ServerKey;
+#[cfg(tarpaulin)]
+use crate::shortint::parameters::coverage_parameters::*;
+use crate::shortint::parameters::*;
+use crate::shortint::PBSParameters;
+
+create_parametrized_test!(integer_smart_neg);
+create_parametrized_test!(integer_default_neg);
+
+fn integer_smart_neg<P>(param: P)
+where
+    P: Into<PBSParameters>,
+{
+    let executor = CpuFunctionExecutor::new(&ServerKey::smart_neg_parallelized);
+    smart_neg_test(param, executor);
+}
+
+fn integer_default_neg<P>(param: P)
+where
+    P: Into<PBSParameters>,
+{
+    let executor = CpuFunctionExecutor::new(&ServerKey::neg_parallelized);
+    default_neg_test(param, executor);
+}
